pyproject.toml
[tool.poetry]
name = "mcp-hello-world"
version = "0.1.0"
description = ""
authors = ["Julian Lopez <jlopezsa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langgraph = "0.6.6"
langchain = "0.3.27"
langchain-openai = "0.3.32"
langchain-mcp-adapters = "0.1.9"
python-dotenv = "^1.2.2"

[tool.poetry.scripts]
mcp-hello-world = "mcp_hello_world.main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
